requirements.txt
python-dotenv>=1.0.1
torch>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
datasets>=2.19.0
transformers>=4.41.0
accelerate>=0.32.0
sentence-transformers>=2.7.0
huggingface-hub>=0.23.0
trafilatura>=1.8.0
llama-index>=0.10.59
llama-index-finetuning>=0.2.1
llama-index-embeddings-huggingface>=0.2.3
llama-index-llms-ollama>=0.1.4
llama-index-readers-web>=0.1.13
llama-index-core>=0.10.59
openai>=1.40.0
mteb>=1.14.15
pandas>=2.2.0
tqdm>=4.66.0

